최종배포안 dev로 이동

+++ server/controller/PlannerController.js
@@ -1,7 +1,24 @@
-const { group, plan, planner } = require('../models')
+const { group, plan, planner } = require('../models');
 const axios  = require('axios');
 
+const { verifyToken, decodeToken } = require('./VerifyToken');
 const SERVER_URL = process.env.SERVER_URL || 'http://localhost:80';
+const ACCESS_COOKIE_OPTIONS = {
+  MaxAge: 1000 * 60 * 60,
+  domain: 'ohjeju.link',
+  path: '/',
+  httpOnly: true,
+  secure: true,
+  sameSite: 'none'
+}
+const REFRESH_COOKIE_OPTIONS = {
+  MaxAge: 1000 * 60 * 60 * 24 * 14,
+  domain: 'ohjeju.link',
+  path: '/',
+  httpOnly: true,
+  secure: true,
+  sameSite: 'none'
+}
 
 //플래너 관련 method
 module.exports = {
@@ -18,7 +35,6 @@ module.exports = {
           })
           .then((planner) => planner.id)
           
-          //클라에서 어떻게 구현되냐에 따라 send 내용 생략할 수 있을 것 같아요. 일단은 필요할 것 같아서 plannerId 리턴했어요
           return res.status(201).send({ plannerId: newPlannerId });
         }
         else { //로그인 상태일 경우 -> 다인 플래너로 생성
@@ -39,22 +55,6 @@ module.exports = {
           return res.status(201).send(newPlanner);
         }
       }
-      
-      // if(req.body.name.length === 0){ //빈칸인지 undefined지는 알아서
-      //   res.status(400).send("content is empty")
-      // }
-      // else{
-      //   if(req.body.groupId) {// 그룹이 생성 되고 플래너 생성할때
-      //     planner.create({
-      //       name:req.body.name,
-      //       groupId:req.body.groupId
-      //     }).then(item => res.status(201).send('ok'))
-      //   }
-      //   else { //플래너 생성 하고 그룹생성할때 -> 이거 어떻게할건지 얘기해봐야됨 순서를 강제할건지
-      //     //정해지면 추후작성
-      //   }
-      // }
-      // return res.status(200).send('test');
     }
     catch(err) { return res.status(500).send('server error') }
   },
@@ -74,31 +74,12 @@ module.exports = {
           return { id, day, departureTime, destination, memo };
         }))
 
-      const customPlanner = (planArr) => {
-        const planner = new Object();
-
-        for(const plan of planArr) {
-          if(!planner[plan.day]) planner[plan.day] = new Array();
-          planner[plan.day].push({
-            id: plan.id,
-            departureTime: plan.departureTime,
-            destination: plan.destination,
-            memo: plan.memo
-          })
-        }
-        for(const day in planner) planner[day].sort((a, b) => a.departureTime - b.departureTime);
-
-        return planner;
-      }
-      // const customedPlanner = customPlanner(planInThis);
-      const customedPlanner = planInThis;
-
       if(!targetPlanner.group) { //1인 플래너일 경우
         return res.status(200).send({
           id:targetPlanner.id,
           name: targetPlanner.name,
           group: null,
-          plan: customedPlanner
+          plan: planInThis
         });
       }
       else { //다인 플래너일 경우
@@ -109,65 +90,9 @@ module.exports = {
           id:targetPlanner.id,
           name: targetPlanner.name,
           group: groupInThis,
-          plan: customedPlanner
+          plan: planInThis
         })
       }
-      // const findPlanner = await planner.findOne({
-      //   where:{id:req.query.plannerId},
-      //   include:[{model:group}]
-      // })
-      // const userGroup = await user_group.findAll({
-      //   where:{groupId:findPlanner.group.id},
-      //   include:[{model:users},{model:group}]
-      // }) //얘는이상하게 axios해서 받아올려하면 오류남 그래서 그냥 sequelize함수로
-      // const userInGroup = userGroup.map(item => item.user.userName)
-      // const findPlan = await plan.findAll({
-      //   where:{plannerId:req.query.plannerId}
-      // })
-      // const customPlan = () => {
-        /* 
-        plan부분 get이 따로 작성한게없고 planner 불러올때 Plan을 불러오기로햇는데
-        api문서대로 불러오게하면 너무 지저분하고 복잡함 plan이 시간대별 일자별로 여러개인데 결과가
-        plan:[{day:1,depart:9시 ..},{day:2,depart:9시 ..}] ...이런식으로 플래너에 일차가 많고 일차당 plan이많을수록 지저분함
-        일단은 틀은
-        plan:{
-          day1:[], 이런식으로나옴 근데 출발시간별 오름차순 이런건 고민좀해봐야할듯
-          dat2:[] -> 안에 day는 중복되는느낌인데 그냥 일단 넣음
-          ..일차 늘어나면 객체key도 늘어남
-        }
-        구상하느라 시간좀썻긴한데 더 불편하면 바까도됨 ㅎ;ㅋ;
-        */
-  
-      //   const resultPlan = {}
-      //   findPlan.forEach(item => {
-      //     const {id,departureTime,destination,memo,day} = item
-      //     const VAR = `day${item.day}`
-      //     if(!resultPlan[VAR]){
-      //       resultPlan[VAR] =[];
-      //       resultPlan[VAR].push({
-      //         id,departureTime,destination,memo,day
-      //       })
-      //     }
-      //     else{
-      //       resultPlan[VAR].push({
-      //         id,departureTime,destination,memo,day
-      //       })
-      //     }
-      //   })
-      //   return resultPlan
-      // }
-      // if(findPlanner === null) res.status(404).send("can't find the planner")
-      // else{
-      //   const resultPlan = customPlan();
-      //   res.json({
-      //     plan:resultPlan,
-      //     group:{
-      //       groupId:findPlanner.group.id,
-      //       userName:userInGroup
-      //     },
-      //     name:findPlanner.name
-      //   })
-      // }
     }
     catch(err) { return res.status(500).send('server error') }
   },
@@ -183,25 +108,6 @@ module.exports = {
         .then((res) => res.data)
 
       return res.status(200).send(modifiedPlanner)
-
-      // const findPlanner = await planner.findOne({
-      //   where:{id:req.query.plannerId},
-      // })
-      // if(findPlanner === null) res.status(404).send("can't find the planner")
-      // else{
-      //   planner.update({
-      //     name:req.body.name
-      //   },{
-      //     where:{id:req.query.plannerId}
-      //   })
-      //   .then(item => {
-      //     //마찬가지
-      //     axios.get(`http://localhost:80/planner?plannerId=${findPlanner.id}`)
-      //     .then(result => {
-      //       res.status(200).json(result.data)
-      //     })
-      //   })
-      // }
     }
     catch(err) { return res.status(500).send('server error') }
   },
@@ -210,9 +116,6 @@ module.exports = {
     //* endpoint: https://ohjeju.link/planner?plannerId=''
 
     try{
-      //그룹 리더일 경우에만 플래너 삭제 가능
-      //1인플래너는? -> 그냥 삭제? / 토큰이 없기때문에 나중에 다시 와서 삭제하는 기능은 구현불가능
-      //일단 코드는 1인플래너는 그냥 삭제하는 쪽으로 작성했어요
       const targetPlanner = await planner.findOne({
         where: { id: req.query.plannerId },
         include: [{ model: group }]
@@ -224,7 +127,13 @@ module.exports = {
         return res.status(200).send('ok');
       }
       else { //다인 플래너일 경우 -> 그룹까지 삭제해야함
-        console.log(targetPlanner.group.id)
+        const [reqAccessToken, reqRefreshToken] = await verifyToken(req);
+        if(!reqAccessToken) return res.status(403).send('can\'t access');
+        const tokenUser = await decodeToken(reqAccessToken);
+        if(!tokenUser) return res.status(403).send('can\'t access');
+
+        //그룹 리더랑 현재 유저랑 같은지 먼저 확인
+        if(tokenUser.id !== targetPlanner.group.leaderId) return res.status(403).send('Only leader can delete planner');
         axios.delete(`${SERVER_URL}/group?groupId=${targetPlanner.group.id}`, {
           headers: {
             cookie: `accessToken=${req.cookies.accessToken};refreshToken=${req.cookies.refreshToken}`
@@ -235,19 +144,11 @@ module.exports = {
             targetPlanner.destroy()
           })
 
-        return res.status(200).send('ok');
+        return res.status(200)
+          .cookie('accessToken', reqAccessToken, ACCESS_COOKIE_OPTIONS)
+          .cookie('refreshToken', reqRefreshToken, REFRESH_COOKIE_OPTIONS)
+          .send('ok');
       }
-
-      // const findPlanner = await planner.findOne({
-      //   where:{id:req.query.plannerId},
-      // })
-      // if(findPlanner === null) res.status(404).send("can't find the planner")
-      // else{
-      //   planner.destroy({
-      //     where:{id:req.query.plannerId}
-      //   })
-      //   .then(item => res.status(200).send('ok'))
-      // }
     }
     catch(err) { return res.status(500).send('server error') }
   }
